refactor(agentic-ai): reorganize classes in the adhoctoolschema package
