+++ connectors/agentic-ai/src/main/java/io/camunda/connector/agenticai/adhoctoolsschema/resolver/feel/FeelInputParamExtractionException.java
@@ -4,7 +4,7 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.agenticai.adhoctoolsschema.feel;
+package io.camunda.connector.agenticai.adhoctoolsschema.resolver.feel;
 
 public class FeelInputParamExtractionException extends RuntimeException {
