+++ connectors/agentic-ai/src/main/java/io/camunda/connector/agenticai/adhoctoolsschema/schema/AdHocToolSchemaGenerator.java
@@ -4,15 +4,15 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.agenticai.adhoctoolsschema.resolver.schema;
+package io.camunda.connector.agenticai.adhoctoolsschema.schema;
 
 import io.camunda.connector.agenticai.adhoctoolsschema.model.AdHocToolElement;
-import io.camunda.connector.agenticai.model.tool.ToolDefinition;
+import java.util.Map;
 
 /**
- * Generates a JSON schema from a list of input parameter definitions resolved from tagging
- * functions such as fromAi().
+ * Generates a JSON schema from an identified ad-hoc tool element including parameter definitions
+ * resolved from tagging functions such as fromAi().
  */
-public interface AdHocToolDefinitionConverter {
-  ToolDefinition createToolDefinition(AdHocToolElement element);
+public interface AdHocToolSchemaGenerator {
+  Map<String, Object> generateToolSchema(AdHocToolElement element);
 }
