+++ connectors/agentic-ai/src/test/java/io/camunda/connector/agenticai/adhoctoolsschema/schema/AdHocToolDefinitionConverterTest.java
@@ -4,10 +4,9 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.agenticai.adhoctoolsschema.resolver.schema;
+package io.camunda.connector.agenticai.adhoctoolsschema.schema;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.when;
 
 import io.camunda.connector.agenticai.adhoctoolsschema.model.AdHocToolElement;
