+++ connectors/agentic-ai/src/test/java/io/camunda/connector/agenticai/adhoctoolsschema/schema/AdHocToolSchemaGeneratorTest.java
@@ -4,7 +4,7 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.agenticai.adhoctoolsschema.resolver.schema;
+package io.camunda.connector.agenticai.adhoctoolsschema.schema;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
